fix to add migration_mode and nil version check

+++ cli/commands/migrate.go
@@ -55,7 +55,7 @@ func ExecuteMigration(cmd string, t *migrate.Migrate, stepOrVersion int64) error
 	case "down":
 		err = mig.DownCmd(t, stepOrVersion)
 	case "gotoVersion":
-		err = mig.GotoVersionCmd(t, uint64(stepOrVersion))
+		err = mig.GotoVersionCmd(t, stepOrVersion)
 	case "version":
 		var direction string
 		if stepOrVersion >= 0 {
